feat: add setup agent for Modius
